+++ example/src/main/java/io/realm/examples/adapters/ui/listview/MyListAdapter.java
@@ -24,15 +24,15 @@ import android.widget.TextView;
 
 import io.realm.OrderedRealmCollection;
 import io.realm.RealmBaseAdapter;
-import io.realm.examples.adapters.model.TimeStamp;
+import io.realm.examples.adapters.model.Counter;
 
-public class MyListAdapter extends RealmBaseAdapter<TimeStamp> implements ListAdapter {
+class MyListAdapter extends RealmBaseAdapter<Counter> implements ListAdapter {
 
     private static class ViewHolder {
-        TextView timestamp;
+        TextView countText;
     }
 
-    public MyListAdapter(OrderedRealmCollection<TimeStamp> realmResults) {
+    MyListAdapter(OrderedRealmCollection<Counter> realmResults) {
         super(realmResults);
     }
 
@@ -43,14 +43,16 @@ public class MyListAdapter extends RealmBaseAdapter<TimeStamp> implements ListAd
             convertView = LayoutInflater.from(parent.getContext())
                     .inflate(android.R.layout.simple_list_item_1, parent, false);
             viewHolder = new ViewHolder();
-            viewHolder.timestamp = (TextView) convertView.findViewById(android.R.id.text1);
+            viewHolder.countText = (TextView) convertView.findViewById(android.R.id.text1);
             convertView.setTag(viewHolder);
         } else {
             viewHolder = (ViewHolder) convertView.getTag();
         }
 
-        TimeStamp item = adapterData.get(position);
-        viewHolder.timestamp.setText(item.getTimeStamp());
+        if (adapterData != null) {
+            Counter item = adapterData.get(position);
+            viewHolder.countText.setText(item.getCountString());
+        }
         return convertView;
     }
 }